[discretizer] base class is now Pipeline
