+++ pyemma/coordinates/util/chaining.py
@@ -3,8 +3,9 @@ Created on 17.02.2015
 
 @author: marscher
 '''
+from pyemma.util.annotators import deprecated
 
-
+@deprecated
 def build_chain(transformers, chunksize=None):
     """
     utility method to build a working pipeline out of given data source and
@@ -20,7 +21,7 @@ def build_chain(transformers, chunksize=None):
 
     return transformers
 
-
+@deprecated
 def run_chain(chain):
     for c in chain:
         if hasattr(c, 'parametrize'):
